create database flag.
In the template some code is locked behind a database flag, to turn this code on a database feature needs to be enabled. We can code the cfg checks to check all 3 database types currently used, or we can use a general feature that is turned on when one of the three database types is enabled. This is basically an empty feature flag. This also allows to add different type of databases in the feature, without having to change the code.